add processing status and stop removing +'s from filenames
The filename is quite a difficult situation. Your best bet is to just
not use filenames with spaces, +'s or parentheses.

If you upload a file to S3 with spaces, the spaces get turned into +'s
in the download URL , but the bucket key remains the version with
spaces.

If you upload a file with +'s, the +'s get turned into %2B in the
download URL, but the bucket key remains the same.

In both cases the downloaded file you end up with on your disk will be
the one with spaces, so you'll have no idea what the key was if you only
have the downloaded file.